packages/migra/pyproject.toml
[project]
name = "migra"
version = "4.0.0"
description = "Like `diff` but for PostgreSQL schemas"
readme = "../../README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Roshan Pawar", email = "roshan@r614.dev" },
]
keywords = ["postgresql", "schema", "diff", "migration", "database"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
]
dependencies = [
    "psycopg[binary]",
    "pyyaml",
]

[project.urls]
Repository = "https://github.com/r614/migra"
Issues = "https://github.com/r614/migra/issues"

[project.scripts]
migra = "migra:do_command"
schemainspect = "migra.schemainspect:do_command"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
